refactoring collect frontier to a better home
